+++ ccxt-exchanges/src/bybit/mod.rs
@@ -179,8 +179,8 @@ impl Bybit {
     }
 
     /// Returns the rate limit in requests per second.
-    pub fn rate_limit(&self) -> f64 {
-        20.0
+    pub fn rate_limit(&self) -> u32 {
+        20
     }
 
     /// Returns `true` if sandbox/testnet mode is enabled.
